cosmosDB connection scripts moved to seperate class from main program class
